+++ src/cljs/cxengage_javascript_sdk/modules/reporting.cljs
@@ -140,18 +140,17 @@
                   :error (e/invalid-args-error (s/explain-data ::get-capacity-params params))
                   :callback callback})
       (go (let [capacity-url (str (st/get-base-api-url) (get-in @module-state [:urls url]))
+                url-params (if resource-id {:tenant-id tenant-id :resource-id resource-id} {:tenant-id tenant-id})
                 capacity-request {:method :get
                                   :url (iu/build-api-url-with-params
                                         capacity-url
-                                        {:tenant-id tenant-id
-                                         :resource-id resource-id})}
+                                        url-params)}
                 {:keys [api-response status]} (a/<! (iu/api-request capacity-request))
-                {:keys [results]} api-response
-                capacity-topic (p/get-topic :get-capacity)]
+                {:keys [results]} api-response]
             (if (not= status 200)
-              (p/publish {:topics capacity-topic
+              (p/publish {:topics topic
                           :error (e/api-error "api returned an error")})
-              (p/publish {:topics capacity-topic
+              (p/publish {:topics topic
                           :response results})))))
     nil)))